[DAD-1061] refactor: board api 코드 리팩토링
